+++ src/base/PluginTrait.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link      https://dukt.net/twitter/
- * @copyright Copyright (c) 2019, Dukt
+ * @copyright Copyright (c) 2021, Dukt
  * @license   https://github.com/dukt/twitter/blob/master/LICENSE.md
  */
 
@@ -30,6 +30,17 @@ trait PluginTrait
     {
         /** @var Twitter $this */
         return $this->get('twitter');
+    }
+
+    /**
+     * Returns the accounts service.
+     *
+     * @return \dukt\twitter\services\Accounts The accounts service
+     */
+    public function getAccounts()
+    {
+        /** @var Twitter $this */
+        return $this->get('accounts');
     }
 
     /**
@@ -77,48 +88,36 @@ trait PluginTrait
     }
 
     /**
-     * Returns the OAuth consumer key.
+     * Gets the OAuth consumer key
      *
+     * @param bool $parse
      * @return string|null
      */
-    public function getConsumerKey()
+    public function getConsumerKey(bool $parse = true)
     {
-        $consumerKey = Plugin::getInstance()->getSettings()->oauthConsumerKey;
+        $consumerKey = Plugin::$plugin->getSettings()->oauthConsumerKey;
 
-        if ($consumerKey) {
-            return $consumerKey;
-        }
-
-        $plugin = Craft::$app->getPlugins()->getPlugin('twitter');
-        $settings = $plugin->getSettings();
-
-        if (empty($settings['oauthConsumerKey'])) {
+        if (!$consumerKey) {
             return null;
         }
 
-        return $settings['oauthConsumerKey'];
+        return $parse ? Craft::parseEnv($consumerKey) : $consumerKey;
     }
 
     /**
-     * Returns the OAuth consumer secret.
+     * Gets the OAuth consumer secret
      *
+     * @param bool $parse
      * @return string|null
      */
-    public function getConsumerSecret()
+    public function getConsumerSecret(bool $parse = true)
     {
-        $consumerSecret = Plugin::getInstance()->getSettings()->oauthConsumerSecret;
+        $consumerSecret = Plugin::$plugin->getSettings()->oauthConsumerSecret;
 
-        if ($consumerSecret) {
-            return $consumerSecret;
-        }
-
-        $plugin = Craft::$app->getPlugins()->getPlugin('twitter');
-        $settings = $plugin->getSettings();
-
-        if (!empty($settings['oauthConsumerSecret'])) {
+        if (!$consumerSecret) {
             return null;
         }
 
-        return $settings['oauthConsumerSecret'];
+        return $parse ? Craft::parseEnv($consumerSecret) : $consumerSecret;
     }
 }