+++ src/controllers/SettingsController.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link      https://dukt.net/twitter/
- * @copyright Copyright (c) 2019, Dukt
+ * @copyright Copyright (c) 2021, Dukt
  * @license   https://github.com/dukt/twitter/blob/master/LICENSE.md
  */
 
@@ -67,12 +67,11 @@ class SettingsController extends Controller
      */
     public function actionOauth(): Response
     {
-        $plugin = Craft::$app->getPlugins()->getPlugin('twitter');
-
         return $this->renderTemplate('twitter/settings/oauth', [
             'javascriptOrigin' => Plugin::getInstance()->getOauth()->getJavascriptOrigin(),
             'redirectUri' => Plugin::getInstance()->getOauth()->getRedirectUri(),
-            'settings' => $plugin->getSettings(),
+            'oauthConsumerKey' => Plugin::$plugin->getConsumerKey(false),
+            'oauthConsumerSecret' => Plugin::$plugin->getConsumerSecret(false),
         ]);
     }