+++ src/migrations/Install.php
@@ -1,0 +1,94 @@
+<?php
+/**
+ * @link      https://dukt.net/twitter/
+ * @copyright Copyright (c) 2021, Dukt
+ * @license   https://github.com/dukt/twitter/blob/master/LICENSE.md
+ */
+
+namespace dukt\twitter\migrations;
+
+use Craft;
+use craft\db\Migration;
+
+class Install extends Migration
+{
+    // Public Properties
+    // =========================================================================
+
+    /**
+     * @var string The database driver to use
+     */
+    public $driver;
+
+    // Public Methods
+    // =========================================================================
+
+    /**
+     * This method contains the logic to be executed when applying this migration.
+     * This method differs from [[up()]] in that the DB logic implemented here will
+     * be enclosed within a DB transaction.
+     * Child classes may implement this method instead of [[up()]] if the DB logic
+     * needs to be within a transaction.
+     *
+     * @return boolean return a false value to indicate the migration fails
+     * and should not proceed further. All other return values mean the migration succeeds.
+     */
+    public function safeUp()
+    {
+        $this->driver = Craft::$app->getConfig()->getDb()->driver;
+        $this->createTables();
+
+        return true;
+    }
+
+    /**
+     * This method contains the logic to be executed when removing this migration.
+     * This method differs from [[down()]] in that the DB logic implemented here will
+     * be enclosed within a DB transaction.
+     * Child classes may implement this method instead of [[down()]] if the DB logic
+     * needs to be within a transaction.
+     *
+     * @return boolean return a false value to indicate the migration fails
+     * and should not proceed further. All other return values mean the migration succeeds.
+     */
+    public function safeDown()
+    {
+        $this->driver = Craft::$app->getConfig()->getDb()->driver;
+        $this->removeTables();
+
+        return true;
+    }
+
+    // Protected Methods
+    // =========================================================================
+
+    /**
+     * Creates the tables needed for the Records used by the plugin
+     *
+     * @return void
+     */
+    protected function createTables()
+    {
+        $this->createTable(
+            '{{%twitter_accounts}}',
+            [
+                'id' => $this->primaryKey(),
+                'token' => $this->string(50),
+                'tokenSecret' => $this->string(50),
+                'dateCreated' => $this->dateTime()->notNull(),
+                'dateUpdated' => $this->dateTime()->notNull(),
+                'uid' => $this->uid()
+            ]
+        );
+    }
+
+    /**
+     * Removes the tables needed for the Records used by the plugin
+     *
+     * @return void
+     */
+    protected function removeTables()
+    {
+        $this->dropTable('{{%twitter_accounts}}');
+    }
+}